# add WEBPACK_VERBOSE (default off), only output some warnings when enabled