--- a/config/webpack/loaders.js
+++ b/config/webpack/loaders.js
@@ -12,14 +12,23 @@ if (env.WEBPACK_EXCLUDE_NODE_MODULES) {
   base.exclude = /node_modules\/d3/;
 }
 
+let babelOptions = merge(babelrc, {
+  babelrc: false,
+  compact: false,
+});
+
+// set WEBPACK_VERBOSE=1 to get more warnings
+if (env.WEBPACK_VERBOSE) {
+  // don't drop any whitespace from the bundle, prevents warnings about filesize over the limit
+  delete babelOptions.compact;
+}
+
 module.exports = [
   merge(base, {
     test: /\.(mjs|js|jsx)$/,
     use: [{
       loader: 'babel-loader',
-      options: merge(babelrc, {
-        babelrc: false,
-      }),
+      options: babelOptions,
     }],
   }),