This removes these kinds of warnings from webpack:

```
[BABEL] Note: The code generator has deoptimised the styling of "/home/dberger/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-ui-classic-615df016338b/node_modules/lodash/lodash.js" as it exceeds the max of "500KB".
```

(coming from babel, when trying to [compact](https://babeljs.io/docs/en/options#compact) big files, removing means a bit larger bundle, but only by whitespace (so presumably compressible))

and
￼
```
WARNING in jquery
  Multiple versions of jquery found:
    2.2.4 ./~/jquery from ./app/javascript/packs/globals.js
    3.3.1 ./~/bootstrap-datepicker/~/jquery from ./~/bootstrap-datepicker/dist/js/bootstrap-datepicker.js
```

(coming from `DuplicatePackageCheckerPlugin` - used to get warned about duplicate deps)

Both can be re-enabled using `WEBPACK_VERBOSE=1 bin/webpack`.