@himdel 2018-12-20
Before, or with `WEBPACK_VERBOSE=1`:

```
yarn run v1.12.3
warning From Yarn 1.0 onwards, scripts don't require "--" for options to be forwarded. In a future version, any explicit "--" will be forwarded as-is to the scripts.
$ /home/himdel/manageiq-ui-classic/node_modules/.bin/webpack --config /home/himdel/manageiq-ui-classic/config/webpack/development.js

[at-loader] Using typescript@2.8.4 from typescript and "tsconfig.json" from /home/himdel/manageiq-ui-classic/tsconfig.json.

[BABEL] Note: The code generator has deoptimised the styling of "/home/himdel/manageiq-ui-classic/node_modules/lodash/lodash.js" as it exceeds the max of "500KB".
[BABEL] Note: The code generator has deoptimised the styling of "/home/himdel/manageiq-ui-classic/node_modules/react-dom/cjs/react-dom.development.js" as it exceeds the max of "500KB".
[BABEL] Note: The code generator has deoptimised the styling of "/home/himdel/manageiq-ui-classic/node_modules/angular/angular.js" as it exceeds the max of "500KB".
[BABEL] Note: The code generator has deoptimised the styling of "/home/himdel/manageiq-ui-classic/node_modules/angular-patternfly/node_modules/angular/angular.js" as it exceeds the max of "500KB".

[at-loader] Checking started in a separate process...

[at-loader] Ok, 6.771 sec.
Hash: 8ad90ffd14f87164abdc
Version: webpack 4.12.2
Time: 112510ms
Built at: 12/20/2018 7:02:42 PM
                                                                        Asset       Size                                                 Chunks             Chunk Names
      manageiq-ui-classic/provisioning-actions-common-ba2907a300eab27969a2.js    2.5 KiB        manageiq-ui-classic/provisioning-actions-common  [emitted]  manageiq-ui-classic/provisioning-actions-common
        manageiq-graphql/component-definitions-common-777dafd08f8db2a25719.js   42.6 KiB          manageiq-graphql/component-definitions-common  [emitted]  manageiq-graphql/component-definitions-common
                   manageiq-ui-classic/angular-common-20ac70b434b594806fc9.js   35.3 KiB                     manageiq-ui-classic/angular-common  [emitted]  manageiq-ui-classic/angular-common
               manageiq-ui-classic/application-common-6edd97a4707043b341fc.js   47.8 KiB                 manageiq-ui-classic/application-common  [emitted]  manageiq-ui-classic/application-common
     manageiq-ui-classic/component-definitions-common-81b503d3b6a9c6c51de0.js    113 KiB       manageiq-ui-classic/component-definitions-common  [emitted]  manageiq-ui-classic/component-definitions-common
                          manageiq-ui-classic/globals-ace665fa83a053ee319f.js   13.2 KiB                            manageiq-ui-classic/globals  [emitted]  manageiq-ui-classic/globals
          manageiq-ui-classic/provider-dialogs-common-66feee5f048a98ea3ed5.js   30.5 KiB            manageiq-ui-classic/provider-dialogs-common  [emitted]  manageiq-ui-classic/provider-dialogs-common
manageiq-providers-nuage/component-definitions-common-213476a322b8adacecfb.js   35.6 KiB  manageiq-providers-nuage/component-definitions-common  [emitted]  manageiq-providers-nuage/component-definitions-common
           manageiq-ui-classic/toolbar-actions-common-cad01519057cade84cff.js   21.4 KiB             manageiq-ui-classic/toolbar-actions-common  [emitted]  manageiq-ui-classic/toolbar-actions-common
                               manageiq-v2v/migration-4931fddb5071c758ce71.js   2.95 MiB                                 manageiq-v2v/migration  [emitted]  manageiq-v2v/migration
                               manageiq-v2v/polyfills-497415e82a49b922b36c.js  466 bytes                                 manageiq-v2v/polyfills  [emitted]  manageiq-v2v/polyfills
                                              runtime-21acf8de622f56dba48a.js   14.2 KiB                                                runtime  [emitted]  runtime
                                                shims-98267b1384cebaad0d7c.js   1.87 MiB                                                  shims  [emitted]  shims
                                               vendor-f7de20330f9838d7884c.js   55.6 MiB                                                 vendor  [emitted]  vendor
                                                                manifest.json   1.46 KiB                                                         [emitted]  
[../.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/app/javascript/graphql-explorer/index.js] 658 bytes {manageiq-graphql/component-definitions-common} [built]
[../.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/app/javascript/packs/component-definitions-common.js] 309 bytes {manageiq-graphql/component-definitions-common} [built]
[../.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/app/javascript/packs/component-definitions-common.js] 389 bytes {manageiq-providers-nuage/component-definitions-common} [built]
[../.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/app/javascript/packs/migration.js] 542 bytes {manageiq-v2v/migration} [built]
[../.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/app/javascript/packs/polyfills.js] 104 bytes {vendor} [built]
[./app/javascript/packs/angular-common.js] 37 bytes {manageiq-ui-classic/angular-common} [built]
[./app/javascript/packs/application-common.js] 2.92 KiB {manageiq-ui-classic/application-common} [built]
[./app/javascript/packs/component-definitions-common.js] 2.75 KiB {manageiq-ui-classic/component-definitions-common} [built]
[./app/javascript/packs/globals.js] 1.57 KiB {manageiq-ui-classic/globals} [built]
[./app/javascript/packs/provider-dialogs-common.js] 5.71 KiB {manageiq-ui-classic/provider-dialogs-common} [built]
[./app/javascript/packs/provisioning-actions-common.js] 451 bytes {manageiq-ui-classic/provisioning-actions-common} [built]
[./app/javascript/packs/toolbar-actions-common.js] 1.49 KiB {manageiq-ui-classic/toolbar-actions-common} [built]
[./app/javascript/toolbar-actions/custom-action.js] 1.07 KiB {manageiq-ui-classic/toolbar-actions-common} [built]
[./app/javascript/toolbar-actions/delete.js] 1.97 KiB {manageiq-ui-classic/toolbar-actions-common} [built]
[5] multi es6-shim array-includes whatwg-fetch babel-polyfill 64 bytes {shims} [built]
    + 3414 hidden modules

WARNING in @babel/runtime
  Multiple versions of @babel/runtime found:
    7.1.5 ./~/dom-helpers/~/@babel/runtime from ./~/dom-helpers/query/position.js
    7.2.0 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/@babel/runtime from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-redux/es/components/connectAdvanced.js


WARNING in @manageiq/react-ui-components
  Multiple versions of @manageiq/react-ui-components found:
    0.10.8 ./~/@manageiq/react-ui-components from ./app/javascript/packs/component-definitions-common.js
    0.9.5 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/@manageiq/react-ui-components from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/app/javascript/components/forms/nuage-cloud-subnet-form.jsx


WARNING in angular
  Multiple versions of angular found:
    1.5.11 ./~/angular-patternfly/~/angular from ./~/angular-patternfly/~/angular/index.js
    1.6.10 ./~/angular from ./~/angular/index.js


WARNING in angular-sanitize
  Multiple versions of angular-sanitize found:
    1.5.11 ./~/angular-patternfly/~/angular-sanitize from ./~/angular-patternfly/~/angular-sanitize/index.js
    1.6.10 ./~/angular-sanitize from ./~/angular-sanitize/index.js


WARNING in angular-ui-bootstrap
  Multiple versions of angular-ui-bootstrap found:
    0.14.3 ./~/angular-patternfly/~/angular-ui-bootstrap from ./~/angular-patternfly/index.js
    2.5.6 ./~/angular-ui-bootstrap from ./~/angular-ui-bootstrap/index.js


WARNING in base64-js
  Multiple versions of base64-js found:
    1.2.3 ./~/base64-js from ./app/javascript/http_api/compat.js
    1.3.0 ./~/buffer/~/base64-js from ./~/buffer/index.js


WARNING in bootstrap-touchspin
  Multiple versions of bootstrap-touchspin found:
    3.1.1 ./~/bootstrap-touchspin from ./app/javascript/packs/globals.js
    4.2.5 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/bootstrap-touchspin from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/app/javascript/react/screens/App/common/forms/NumberInput.js


WARNING in dom-helpers
  Multiple versions of dom-helpers found:
    3.3.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/dom-helpers from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-bootstrap/es/Dropdown.js
    3.4.0 ./~/dom-helpers from ./~/react-bootstrap/es/Dropdown.js


WARNING in graphql-language-service-parser
  Multiple versions of graphql-language-service-parser found:
    0.1.14 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/~/graphql-language-service-parser from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/~/graphql-language-service-parser/dist/index.js
    1.2.2 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/~/graphql-language-service-interface/~/graphql-language-service-parser from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/~/graphql-language-service-interface/~/graphql-language-service-parser/dist/index.js


WARNING in hoist-non-react-statics
  Multiple versions of hoist-non-react-statics found:
    2.5.5 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/hoist-non-react-statics from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-redux/es/components/connectAdvanced.js
    3.2.0 ./~/react-redux/~/hoist-non-react-statics from ./~/react-redux/es/components/connectAdvanced.js
    3.2.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-redux/~/hoist-non-react-statics from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-redux/es/components/connectAdvanced.js


WARNING in isarray
  Multiple versions of isarray found:
    0.0.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/path-to-regexp/~/isarray from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/path-to-regexp/index.js
    1.0.0 ./~/isarray from ./~/readable-stream/lib/_stream_readable.js


WARNING in jquery
  Multiple versions of jquery found:
    2.2.4 ./~/jquery from ./app/javascript/packs/globals.js
    3.3.1 ./~/bootstrap-datepicker/~/jquery from ./~/bootstrap-datepicker/dist/js/bootstrap-datepicker.js


WARNING in lodash
  Multiple versions of lodash found:
    3.10.1 ./~/angular-patternfly/~/lodash from ./~/angular-patternfly/index.js
    4.17.11 ./~/lodash from ./~/lodash/_getTag.js


WARNING in patternfly
  Multiple versions of patternfly found:
    3.31.2 ./~/angular-patternfly/~/patternfly from ./~/angular-patternfly/index.js
    3.54.8 ./~/patternfly from ./app/javascript/packs/globals.js
    3.58.0 ./~/@manageiq/react-ui-components/~/patternfly from ./~/@manageiq/react-ui-components/~/patternfly-react/dist/esm/common/patternfly.js


WARNING in patternfly-react
  Multiple versions of patternfly-react found:
    2.10.1 ./~/patternfly-react from ./~/patternfly-react/dist/esm/components/VerticalNav/VerticalNav.js
    2.25.3 ./~/@manageiq/react-ui-components/~/patternfly-react from ./~/@manageiq/react-ui-components/~/patternfly-react/dist/esm/common/helpers.js


WARNING in performance-now
  Multiple versions of performance-now found:
    0.2.0 ./~/react-motion/~/performance-now from ./~/react-motion/lib/Motion.js
    2.1.0 ./~/performance-now from ./~/raf/index.js


WARNING in react-bootstrap
  Multiple versions of react-bootstrap found:
    0.32.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-bootstrap from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-bootstrap/es/index.js
    0.32.4 ./~/react-bootstrap from ./~/react-bootstrap/es/index.js


WARNING in react-is
  Multiple versions of react-is found:
    16.4.2 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-is from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-is/index.js
    16.6.3 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-is from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-is/index.js


WARNING in react-redux
  Multiple versions of react-redux found:
    5.0.7 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-redux from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-redux/es/index.js
    5.1.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-redux from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-redux/es/index.js


WARNING in react-transition-group
  Multiple versions of react-transition-group found:
    2.4.0 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-transition-group from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-bootstrap/es/Fade.js
    2.5.0 ./~/react-transition-group from ./~/react-bootstrap/es/Fade.js


WARNING in redux
  Multiple versions of redux found:
    3.7.2 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/redux from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/redux/es/index.js
    4.0.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/redux from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-providers-nuage-86289b516806/~/react-redux/es/connect/mapDispatchToProps.js


WARNING in uncontrollable
  Multiple versions of uncontrollable found:
    4.1.0 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/uncontrollable from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/uncontrollable/index.js
    5.1.0 ./~/uncontrollable from ./~/react-bootstrap/es/TabContainer.js


WARNING in warning
  Multiple versions of warning found:
    3.0.0 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/warning from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/history/es/createMemoryHistory.js
    4.0.1 /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-router-dom/~/warning from /home/himdel/.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-v2v-46f03211dbed/~/react-router-dom/es/BrowserRouter.js
    4.0.2 ./~/react-bootstrap-typeahead/~/warning from ./~/react-bootstrap-typeahead/lib/utils/warn.js

Done in 113.25s.
```

After:

```
yarn run v1.12.3
warning From Yarn 1.0 onwards, scripts don't require "--" for options to be forwarded. In a future version, any explicit "--" will be forwarded as-is to the scripts.
$ /home/himdel/manageiq-ui-classic/node_modules/.bin/webpack --config /home/himdel/manageiq-ui-classic/config/webpack/development.js

[at-loader] Using typescript@2.8.4 from typescript and "tsconfig.json" from /home/himdel/manageiq-ui-classic/tsconfig.json.


[at-loader] Checking started in a separate process...

[at-loader] Ok, 6.308 sec.
Hash: 2586e10a1a449bb13dc5
Version: webpack 4.12.2
Time: 104974ms
Built at: 12/20/2018 7:00:03 PM
                                                                        Asset       Size                                                 Chunks             Chunk Names
      manageiq-ui-classic/provisioning-actions-common-ba2907a300eab27969a2.js    2.5 KiB        manageiq-ui-classic/provisioning-actions-common  [emitted]  manageiq-ui-classic/provisioning-actions-common
        manageiq-graphql/component-definitions-common-777dafd08f8db2a25719.js   42.6 KiB          manageiq-graphql/component-definitions-common  [emitted]  manageiq-graphql/component-definitions-common
                   manageiq-ui-classic/angular-common-20ac70b434b594806fc9.js   35.3 KiB                     manageiq-ui-classic/angular-common  [emitted]  manageiq-ui-classic/angular-common
               manageiq-ui-classic/application-common-6edd97a4707043b341fc.js   47.8 KiB                 manageiq-ui-classic/application-common  [emitted]  manageiq-ui-classic/application-common
     manageiq-ui-classic/component-definitions-common-81b503d3b6a9c6c51de0.js    113 KiB       manageiq-ui-classic/component-definitions-common  [emitted]  manageiq-ui-classic/component-definitions-common
                          manageiq-ui-classic/globals-ace665fa83a053ee319f.js   13.2 KiB                            manageiq-ui-classic/globals  [emitted]  manageiq-ui-classic/globals
          manageiq-ui-classic/provider-dialogs-common-66feee5f048a98ea3ed5.js   30.5 KiB            manageiq-ui-classic/provider-dialogs-common  [emitted]  manageiq-ui-classic/provider-dialogs-common
manageiq-providers-nuage/component-definitions-common-213476a322b8adacecfb.js   35.6 KiB  manageiq-providers-nuage/component-definitions-common  [emitted]  manageiq-providers-nuage/component-definitions-common
           manageiq-ui-classic/toolbar-actions-common-cad01519057cade84cff.js   21.4 KiB             manageiq-ui-classic/toolbar-actions-common  [emitted]  manageiq-ui-classic/toolbar-actions-common
                               manageiq-v2v/migration-4931fddb5071c758ce71.js   2.95 MiB                                 manageiq-v2v/migration  [emitted]  manageiq-v2v/migration
                               manageiq-v2v/polyfills-497415e82a49b922b36c.js  466 bytes                                 manageiq-v2v/polyfills  [emitted]  manageiq-v2v/polyfills
                                              runtime-21acf8de622f56dba48a.js   14.2 KiB                                                runtime  [emitted]  runtime
                                                shims-98267b1384cebaad0d7c.js   1.87 MiB                                                  shims  [emitted]  shims
                                               vendor-9b6997423cd97a912a5d.js   56.3 MiB                                                 vendor  [emitted]  vendor
                                                                manifest.json   1.46 KiB                                                         [emitted]  
[../.rbenv/versions/2.4.3/lib/ruby/gems/2.4.0/bundler/gems/manageiq-graphql-ca61cf3d9baf/app/javascript/graphql-explorer/index.js] 658 bytes {manageiq-graphql/component-definitions-common} [built]
…